Extract newsletter validation to a service
[MAILPOET-4236]
